requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6


pytesseract==0.3.10
easyocr==1.7.0
Pillow==9.5.0
pdf2image==1.17.0
pypdf2==3.0.1


spacy==3.7.2
regex==2023.10.3
python-dateutil==2.8.2


scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2


sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
supabase==2.0.3


python-dotenv==1.0.0
